modify scripts logic & and add demos notebook
