+++ .gitignore
@@ -1,10 +1,11 @@
 # train data
 use_case_sport_classification/data/train/
 use_case_sport_classification/data/valid/
+use_case_sport_classification/data/test/
 
 # models
-use_case_sport_classification/models/saved_models/FineTunedEfficientNet_21_06.pth
-
+use_case_sport_classification/models_weight/FineTunedEfficientNet_30epochs.pth
+use_case_sport_classification/models_weight/VGGFineTuned.pth
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
