+++ use_case_sport_classification/README.md
@@ -1,0 +1,29 @@
+# HES-XPLAIN Notebook Repository
+
+Welcome to the HES-XPLAIN Notebook Repository! 📓
+
+## Purpose
+This collection features demo notebooks showcasing eXplainable AI (XAI) methods. These notebooks demonstrate how to implement and utilize XAI techniques to enhance understanding within machine learning models.
+
+## Contents
+### 1. Demo Notebooks
+Explore various notebooks illustrating practical uses of XAI techniques in different machine learning scenarios.
+
+### 2. Sports Image Classification Use Case
+#### Objective:
+This notebook focuses on sports image classification. It utilizes a Convolutional Neural Network (CNN) model and demonstrates the integration of XAI techniques for greater interpretability.
+
+#### Features:
+- **Data:** Sports image dataset
+- **Model:** CNN architecture
+- **XAI Techniques:** Demonstrates how XAI methods can enhance insights into the CNN model's decision-making process.
+
+## How to Use
+1. **Clone** the repository to your machine.
+2. **Explore** demo notebooks to understand XAI techniques.
+3. **Run** the sports image classification notebook to see XAI in action for better model interpretability.
+
+## Contributions
+Your contributions and feedback are welcome! Feel free to suggest improvements, submit new XAI techniques, or raise issues via pull requests or discussions.
+
+Discover the power of eXplainable AI and demystify machine learning decisions!
