--- use_case_sport_classification/data/class_dict.json
@@ -1,1 +1,0 @@
-{"0": "airhockey", "1": "amputefootball", "2": "archery", "3": "armwrestling", "4": "axethrowing", "5": "balancebeam", "6": "barellracing", "7": "baseball", "8": "basketball", "9": "batontwirling", "10": "bikepolo", "11": "billiards", "12": "bmx", "13": "bobsled", "14": "bowling", "15": "boxing", "16": "bullriding", "17": "bungeejumping", "18": "canoeslamon", "19": "cheerleading", "20": "chuckwagonracing", "21": "cricket", "22": "croquet", "23": "curling", "24": "discgolf", "25": "fencing", "26": "fieldhockey", "27": "figureskatingmen", "28": "figureskatingpairs", "29": "figureskatingwomen", "30": "flyfishing", "31": "football", "32": "formula1racing", "33": "frisbee", "34": "gaga", "35": "giantslalom", "36": "golf", "37": "hammerthrow", "38": "hanggliding", "39": "harnessracing", "40": "highjump", "41": "hockey", "42": "horsejumping", "43": "horseracing", "44": "horseshoepitching", "45": "hurdles", "46": "hydroplaneracing", "47": "iceclimbing", "48": "iceyachting", "49": "jaialai", "50": "javelin", "51": "jousting", "52": "judo", "53": "lacrosse", "54": "logrolling", "55": "luge", "56": "motorcycleracing", "57": "mushing", "58": "nascarracing", "59": "olympicwrestling", "60": "parallelbar", "61": "poleclimbing", "62": "poledancing", "63": "polevault", "64": "polo", "65": "pommelhorse", "66": "rings", "67": "rockclimbing", "68": "rollerbladeracing", "69": "rollerderby", "70": "rowing", "71": "rugby", "72": "sailboatracing", "73": "shotput", "74": "shuffleboard", "75": "sidecarracing", "76": "skijumping", "77": "skydiving", "78": "skysurfing", "79": "snowboarding", "80": "snowmobileracing", "81": "speedskating", "82": "steerwrestling", "83": "sumowrestling", "84": "surfing", "85": "swimming", "86": "tabletennis", "87": "tennis", "88": "trackbicycle", "89": "trapeze", "90": "tugofwar", "91": "ultimate", "92": "unevenbars", "93": "volleyball", "94": "watercycling", "95": "waterpolo", "96": "weightlifting", "97": "wheelchairbasketball", "98": "wheelchairracing", "99": "wingsuitflying"}
